Disable `alces sync` for `root`.

+++ etc/profile.d/70-flight-sync.sh
@@ -4,7 +4,7 @@
 ## Copyright (c) 2016 Alces Software Ltd
 ##
 ################################################################################
-if [ ! -f "${XDG_CONFIG_HOME:-$HOME/.config}/clusterware/sync.rc" ]; then
+if [ "$UID" != "0" -a ! -f "${XDG_CONFIG_HOME:-$HOME/.config}/clusterware/sync.rc" ]; then
     mkdir -p "${XDG_CONFIG_HOME:-$HOME/.config}/clusterware"
     cat <<EOF > "${XDG_CONFIG_HOME:-$HOME/.config}/clusterware/sync.rc"
 ################################################################################
